Move workflow definition hydration into hub-sessions
hydrateDefinition reads a workflow asset's workflow.json and validates
it against the envelope schema. Its only dependencies -- the asset
service, the envelope schema, and the workflow.json path constant -- all
live in hub-sessions, so hub-api was never its natural home. Deploy code
in hub-sessions now needs the same read-and-validate, but the dependency
edge runs hub-api to hub-sessions and cannot be reversed. Promote the
function to hub-sessions beside the schema it validates against; hub-api
imports it back at its three call sites. No behavior change.
