+++ packages/hub-sessions/src/workflow-kind.ts
@@ -25,6 +25,7 @@ import { type } from "arktype";
 import { getLogger } from "@intx/log";
 import { CredentialBinding, GrantRequirement } from "@intx/types";
 import { glob, repoActionToGrantVerb } from "@intx/hub-common";
+import type { WorkflowDefinition } from "@intx/workflow/definition";
 import {
   UserPrincipal,
   type AuthorizeFn,
@@ -32,6 +33,7 @@ import {
   type Principal,
   type ValidatePushResult,
 } from "./repo-store";
+import type { AssetService } from "./asset-service";
 
 const logger = getLogger(["hub-sessions", "workflow-kind"]);
 
@@ -108,6 +110,41 @@ export const workflowDefinitionEnvelopeSchema = type({
   // passed through to launch-time resolution unchecked.
   "credentialBindings?": CredentialBinding.array(),
 }).onUndeclaredKey("ignore");
+
+/**
+ * Read and hydrate the workflow definition from a workflow asset's
+ * `workflow.json`. Validates the structural envelope at this boundary,
+ * mirroring the workflow-host child's `loadWorkflowDefinition`: the
+ * per-primitive narrows live in the runtime layer that consumes the
+ * definition, so the envelope check plus the documented narrow is the
+ * canonical hydration shape.
+ */
+export async function hydrateDefinition(
+  assetService: AssetService,
+  assetId: string,
+): Promise<WorkflowDefinition> {
+  const raw = await assetService.readAssetBlob({
+    assetId,
+    path: WORKFLOW_JSON_PATH,
+  });
+  let parsed: unknown;
+  try {
+    parsed = JSON.parse(new TextDecoder().decode(raw));
+  } catch (cause) {
+    throw new Error(
+      `workflow asset ${assetId} ${WORKFLOW_JSON_PATH} is not valid JSON`,
+      { cause },
+    );
+  }
+  const validated = workflowDefinitionEnvelopeSchema(parsed);
+  if (validated instanceof type.errors) {
+    throw new Error(
+      `workflow asset ${assetId} ${WORKFLOW_JSON_PATH} failed envelope validation: ${validated.summary}`,
+    );
+  }
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-type-assertion -- envelope schema enforces structural shape; per-primitive narrows live in the runtime layer that consumes the definition, matching loadWorkflowDefinition in @intx/workflow-host
+  return validated as unknown as WorkflowDefinition;
+}
 
 /**
  * Capability-declarations.json is held to "is a JSON object" at this
